// Frontend/my-app/src/model/OrderItem.d.ts
export interface OrderItem {
  id: number;
  orderId: number;
  quantity: number;
  productId: number;
  feedBack: string;
  voteStar: number;
  firstname: string;
  lastname: string;
  productName: string;
  author: string;
  price: number;
  orderDate: Date;
  path: string;
  description:string;
}